+++ main.lua
@@ -1,5 +1,6 @@
 dofile("./options.lua")
 dofile("./lua_extras/tables.lua")
+refreshItemFlags()
 dofile("./lua_extras/logic.lua")
 
 boxWidth = 7
@@ -47,15 +48,14 @@ local LocationValues = refreshRawLocationValues()
 
 
 while true do
-	moduloRefreshItems = emu.framecount() % items_refreshRate
-	moduloRefreshLocations = emu.framecount() % locations_refreshRate
+	moduloRefresh = emu.framecount() % refreshRate
 	readTitleScreen = readRam("titleScreen")
 	framecount_old = framecount + 1
 	framecount = emu.framecount()
 	framecount_difference = math.abs(framecount - framecount_old)
 	
 	if readTitleScreen == 0 then
-		if moduloRefreshItems == 0 then
+		if moduloRefresh == 0 then
 			if item_tracker == true then
 				forms.clear(picture_box_items, backgroundColor)
 				forms.drawText(picture_box_items, 30, 200, "Load save file!", "#FFFFFF", "#111111", 36, nil, "bold") 
@@ -67,31 +67,27 @@ while true do
 			end
 		end 
 	else
-		if item_tracker == true then
-			if (moduloRefreshItems == 0) or (framecount_difference > 7) then
-				dofile("./options.lua")
-				ItemValues_old = ItemValues
-				ItemValues = refreshRawItemValues()
-				if ItemValues_old ~= ItemValues then
-					forms.clear(picture_box_items, backgroundColor)
-					refreshItems()	
-					forms.refresh(picture_box_items)
-				end
+		if (moduloRefresh == 0) or (framecount_difference > 7) then
+			dofile("./options.lua")
+			ItemValues_old = ItemValues
+			ItemValues = refreshRawItemValues()
+			booleanItem = ItemValues_old ~= ItemValues
+			if item_tracker == true and booleanItem then
+				forms.clear(picture_box_items, backgroundColor)
+				refreshItems()	
+				forms.refresh(picture_box_items)
 			end
-		end
-		if location_tracker == true then
-			if (moduloRefreshLocations == 0) then
-				dofile("./options.lua")
+			
+			if location_tracker == true then
 				LocationValues_old = LocationValues
 				LocationValues = refreshRawLocationValues()
-				if LocationValues_old ~= LocationValues then
+				if (LocationValues_old ~= LocationValues) or booleanItem then
 					forms.clear(picture_box_locations, backgroundColor)
 					loadLocFlags()
 					forms.refresh(picture_box_locations)
 				end
 			end
 		end
-				
 	end
 	emu.frameadvance()
 end